+++ server/src/main/java/org/apache/kafka/server/partition/AlterPartitionListener.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.storage.internals.log;
+package org.apache.kafka.server.partition;
 
 /**
  * A listener that is invoked when the ISR of a partition is altered. Implementations of this
