+++ server/src/main/java/org/apache/kafka/server/partition/AssignmentState.java
@@ -14,30 +14,34 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.storage.internals.log;
+package org.apache.kafka.server.partition;
 
 import java.util.List;
 
-public record OngoingReassignmentState(
-        List<Integer> addingReplicas,
-        List<Integer> removingReplicas,
-        List<Integer> replicas
-) implements AssignmentState {
+/**
+ * Represents the current assignment of replicas for a partition. This can be a simple assignment
+ * or an ongoing reassignment.
+ */
+public interface AssignmentState {
 
-    public OngoingReassignmentState {
-        addingReplicas = List.copyOf(addingReplicas);
-        removingReplicas = List.copyOf(removingReplicas);
-        replicas = List.copyOf(replicas);
-    }
+    /**
+     * An ordered sequence of all the broker ids that were assigned to this topic partition.
+     * @return the list of broker ids
+     */
+    List<Integer> replicas();
 
-    @Override
-    public int replicationFactor() {
-        // Keep the size of the original replicas. Replicas may also include those currently being added.
-        return (int) replicas.stream().filter(r -> !addingReplicas.contains(r)).count();
-    }
+    /**
+     * The number of replicas in the assignment.
+     * @return the replication factor
+     */
+    int replicationFactor();
 
-    @Override
-    public boolean isAddingReplica(int brokerId) {
-        return addingReplicas.contains(brokerId);
-    }
+    /**
+     * Check whether a replica is being added to the assignment.
+     * The simple assignment returns false permanently.
+     *
+     * @param brokerId the broker id to check
+     * @return true if the broker is being added
+     */
+    boolean isAddingReplica(int brokerId);
 }
